server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.6.1",
    "deepgram-sdk>=5.2.0",
    "dotenv>=0.9.9",
    "fastapi>=0.120.3",
    "gtts>=2.5.4",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.1",
    "librosa>=0.11.0",
    "manim>=0.19.0",
    "moviepy>=2.2.1",
    "pydantic>=2.12.3",
    "returns>=0.26.0",
    "ruff>=0.14.2",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
    "yt-dlp>=2025.10.22",
]
